Update configuration and remove fixtures
