+++ tests/integration/test_vulnerability_detector/conftest.py
@@ -6,14 +6,10 @@ import os
 import time
 import pytest
 
-import wazuh_testing.vulnerability_detector
 from wazuh_testing.tools import LOG_FILE_PATH
 from wazuh_testing.tools.file import truncate_file
 from wazuh_testing.tools.monitoring import FileMonitor
 from wazuh_testing.tools.services import control_service
-from wazuh_testing.vulnerability_detector import  delete_vulnerability, update_last_scan, modify_system
-from wazuh_testing.vulnerability_detector import insert_vulnerability, insert_package, clean_table
-from wazuh_testing.vulnerability_detector import DEFAULT_PACKAGE_NAME, DEFAULT_VULNERABILITY_ID, PACKAGES_DB_PATH
 
 @pytest.fixture(scope='module')
 def restart_modulesd(get_configuration, request):
@@ -26,22 +22,3 @@ def restart_modulesd(get_configuration, request):
         control_service('start', daemon='wazuh-modulesd')
     except:
       pass
-
-
-@pytest.fixture(scope='module')
-def generate_default_alert():
-    control_service('stop', daemon='wazuh-db')
-    time.sleep(10)
-    clean_table(os.path.join(PACKAGES_DB_PATH,f"000.db"), 'sys_programs')
-    delete_vulnerability(DEFAULT_VULNERABILITY_ID)
-    insert_package()
-    insert_vulnerability()
-    modify_system()
-    control_service('start', daemon='wazuh-db')
-
-
-@pytest.fixture(scope='module')
-def reset_last_scan():
-    control_service('stop', daemon='wazuh-db')
-    update_last_scan()
-    control_service('start', daemon='wazuh-db')
