+++ tests/integration/test_vulnerability_detector/test_general_settings/data/wazuh_ignore_time.yaml
@@ -9,7 +9,7 @@
     - enabled:
         value: 'yes'
     - run_on_start:
-        value: 'no'
+        value: 'yes'
     - interval:
         value: INTERVAL
     - ignore_time:
